feat: route OTP login through backend instead of Auth Proxy
- Update registerWithOTP to use /auth/init/otp endpoint
- Update loginWithOTP to send verificationToken and clientSignature to backend
- Add buildClientSignature utility for signing OTP login requests
- Remove otpLogin from authProxy client (now handled by backend)
- Remove subOrganizationId from verifyOTP parameters (backend handles it)
