+++ packages/core/src/client/index.ts
@@ -1,14 +1,9 @@
 export {
   type AuthProxyClient,
   type AuthProxyClientConfig,
-  type AuthProxyOtpLoginRequest,
-  type AuthProxyOtpLoginResponse,
   type AuthProxyVerifyOtpRequest,
   type AuthProxyVerifyOtpResponse,
-  type ClientSignature,
   createAuthProxyClient,
-  type OtpLoginParams,
-  type SignaturePayload,
 } from './authProxy.js'
 export {
   createBaseClient,
